// frontend/web/src/services/API/proserve-wb-packaging-api/models/Component.ts
export interface Component {
    /**
     * Component ID
     * @type {string}
     * @memberof Component
     */
    componentId: string;
    /**
     * Component Description
     * @type {string}
     * @memberof Component
     */
    componentDescription: string;
    /**
     * Component Name
     * @type {string}
     * @memberof Component
     */
    componentName: string;
    /**
     * Operating system architectures that are compatible/supported by the component
     * @type {Array<string>}
     * @memberof Component
     */
    componentSupportedArchitectures: Array<string>;
    /**
     * Component Platform
     * @type {string}
     * @memberof Component
     */
    componentPlatform: string;
    /**
     * Component Supported OS Versions
     * @type {Array<string>}
     * @memberof Component
     */
    componentSupportedOsVersions: Array<string>;
    /**
     * Status of the component (CREATED/ARCHIVED)
     * @type {string}
     * @memberof Component
     */
    status: string;
    /**
     * Create Date
     * @type {string}
     * @memberof Component
     */
    createDate: string;
    /**
     * Created By
     * @type {string}
     * @memberof Component
     */
    createdBy: string;
    /**
     * Last Update Date
     * @type {string}
     * @memberof Component
     */
    lastUpdateDate: string;
    /**
     * Last Update By
     * @type {string}
     * @memberof Component
     */
    lastUpdatedBy: string;
}